+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Changes
+
+### Fixed
+- Pagination: always display pagination for multiple pages dataset
+
 ## [0.19.3] - 2020-07-20
 
 ### Fixed